refactor: move new event extension modules into events subpackages
