+++ axis/models/events/__init__.py
@@ -1,0 +1,1 @@
+"""Event-related models namespace."""
